refactoring layout; refactoring scene-builder folder structure
