--- src/layout/components/theme-switcher/index.tsx
@@ -1,22 +1,0 @@
-import { isMobile } from 'react-device-detect';
-
-import Icon from 'components/custom/icon';
-import { useGeneral } from 'providers/general-provider';
-
-import s from './s.module.scss';
-
-const ThemeSwitcher: React.FC = () => {
-  const { toggleDarkTheme, isDarkTheme } = useGeneral();
-
-  if (isMobile) {
-    return null;
-  }
-
-  return (
-    <button type="button" className={s.button} onClick={toggleDarkTheme}>
-      <Icon name={isDarkTheme ? 'theme-switcher-sun' : 'theme-switcher-moon'} width={24} height={24} />
-    </button>
-  );
-};
-
-export default ThemeSwitcher;
