--- src/layout/components/theme-switcher/s.module.scss
@@ -1,20 +1,0 @@
-.button {
-  position: fixed;
-  bottom: 10px;
-  left: 10px;
-  background: rgba(51, 51, 51, 0.6);
-  border: 0;
-  padding: 0;
-  border-radius: 12px;
-  width: 46px;
-  height: 46px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  opacity: 0.4;
-  color: #fff;
-}
-
-.button:hover {
-  opacity: 1;
-}
