+++ src/layout/layout.module.scss
@@ -1,4 +1,4 @@
-.layout {
+.root {
   background: var(--theme-body-color);
   display: flex;
   min-height: 100vh;
