+++ src/modules/land-works/index.tsx
@@ -5,16 +5,10 @@ import { Route, Switch } from 'react-router-dom';
 import ProtectedRoute from 'components/custom/protected-route';
 import { useWarning } from 'providers/warning-provider';
 
-import SceneBuilderFormView from './views/scene-builder-form-view';
-import SceneBuilderView from './views/scene-builder-view';
-import SingleBuilderView from './views/single-scene-builder-view';
-
-const LandsView = lazy(() => import('./views/lands-view'));
 const ExploreView = lazy(() => import('./views/explore-view'));
 const MyPropertiesView = lazy(() => import('./views/my-properties-view'));
 const SingleLand = lazy(() => import('./views/single-land-view'));
 const EditProperty = lazy(() => import('./views/edit-property-view'));
-const LandingView = lazy(() => import('modules/landing'));
 
 const LandworksView: React.FC = () => {
   const warning = useWarning();
@@ -24,7 +18,7 @@ const LandworksView: React.FC = () => {
 
     if (isMobile) {
       warningDestructor = warning.addWarn({
-        text: 'Transactions can only be made from the desktop version using Metamask',
+        text: 'Transactions can only be made from the desktop version using a wallet',
         closable: true,
         storageIdentity: 'bb_desktop_metamask_tx_warn',
       });
@@ -38,12 +32,8 @@ const LandworksView: React.FC = () => {
   return (
     <Switch>
       <Route path="/property/:tokenId" exact component={SingleLand} />
-      <ProtectedRoute isAuthenticated={false} authenticationPath="/explore" path="/all" exact component={LandsView} />
       <Route path="/explore" exact component={ExploreView} />
       <Route path="/my-properties" exact component={MyPropertiesView} />
-      <Route path="/scene-builder" exact component={SceneBuilderView} />
-      <Route path="/scene-builder/:builderName" exact component={SingleBuilderView} />
-      <Route path="/join-builders" exact component={SceneBuilderFormView} />
       <ProtectedRoute
         isAuthenticated={false}
         authenticationPath="/explore"
@@ -51,7 +41,6 @@ const LandworksView: React.FC = () => {
         exact
         component={EditProperty}
       />
-      <Route path={['/home', '/']} component={LandingView} />
     </Switch>
   );
 };
