--- src/modules/land-works/views/lands-view/index.scss
@@ -1,123 +1,0 @@
-@import '../../../../styles/text.scss';
-
-.lands-container {
-  .filters {
-    margin-bottom: 40px;
-    display: flex;
-    align-items: center;
-
-    @media all and (max-width: 645px) {
-      margin-bottom: 20px;
-    }
-
-    > div {
-      @media all and (max-width: 645px) {
-        width: 100%;
-        margin-bottom: 15px;
-
-        > div {
-          width: 100%;
-          text-align: center;
-          span:nth-of-type(2) {
-            padding-right: 0;
-          }
-        }
-      }
-    }
-  }
-
-  .actions-container {
-    margin-bottom: 30px;
-  }
-}
-
-.list-new-property {
-  font-size: 14px;
-  font-weight: 600;
-  text-transform: uppercase;
-  box-shadow: 0px 0px 1px 1px var(--theme-neon-pink-color);
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-
-  .add-icon {
-    margin-right: 10px;
-  }
-
-  @media (max-width: 768px) {
-    display: none;
-  }
-}
-
-.addTokenButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 0;
-  background: transparent;
-  padding: 0;
-
-  @media (max-width: 768px) {
-    display: none;
-  }
-
-  svg {
-    width: 36px;
-    height: 36px;
-    transform: scale(2);
-  }
-}
-
-.lands-pagination {
-  margin-top: 60px;
-
-  .ant-pagination-options-size-changer {
-    position: absolute;
-    right: 0;
-    top: 0;
-    .ant-select-selector {
-      background: none;
-    }
-
-    .ant-select-arrow {
-      color: var(--theme-primary-color);
-    }
-  }
-
-  .ant-pagination-item,
-  .ant-pagination-item-link,
-  .ant-select-selector {
-    color: var(--theme-primary-color);
-    border-radius: 10px;
-    a {
-      color: var(--theme-primary-color);
-    }
-    font-weight: 700;
-    background: none;
-
-    &:hover {
-      border: 1px solid #e255cf;
-    }
-  }
-
-  .ant-select-selector {
-    border: 1px solid var(--theme-border-color) !important;
-  }
-
-  .ant-pagination-item-active {
-    background: none;
-    border: 1px solid #e255cf;
-    a {
-      color: var(--theme-primary-color) !important;
-    }
-  }
-
-  .ant-pagination-options {
-    display: inline-block;
-  }
-}
-
-.lands-claim-container {
-  justify-content: flex-end;
-  display: flex;
-}
