+++ src/modules/scene-builder/components/land-works-scene-builders-heading/assets/scene-expert-landing.png
