+++ src/modules/scene-builder/components/scene-builder-card/card-loader-skeleton/index.tsx
