+++ src/modules/scene-builder/components/scene-builder-card/card-loader-skeleton/styled.tsx
