+++ src/modules/scene-builder/components/scene-builder-card/styled.tsx
