+++ src/modules/scene-builder/components/scene-builder-profile/card-loader-skeleton/index.tsx
