+++ src/modules/scene-builder/components/scene-builder-profile/styled.tsx
