+++ src/modules/scene-builder/components/scene-builder-tabs/index.tsx
