+++ src/modules/scene-builder/components/scene-builder-tabs/styled.tsx
